+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
   - Removed unreliable deepdiff rounding
   - Removed soft Correlation Coefficient
   - Custom accuracy none rounding driven custom diff check
+- Added candles_by_indicator method to find correct Candles list in Hexital
 - Added Movements
   - Added flipped, to detect changes in bool's or solid values
 - Added Indicators